// src/utils/interfaces/searchresults.interface.ts
export interface ISearchResult {
    companyName?: string;
    address?: string;
    city?: string;
    state?: string;
    country?: string;
    pincode?: number;
    email?: string;
}

export interface IFormData {
    companyName?: string;
    address?: string;
    city?: string;
    state?: string;
    country?: string;
    pincode?: number;
    email?: string;
}